backend/pyproject.toml
[tool.poetry]
name = "evf-portugal-2030"
version = "1.0.0"
description = "AI-powered B2B SaaS for Portuguese funding applications"
authors = ["EVF Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.35"}
asyncpg = "^0.29.0"
alembic = "^1.13.2"
redis = "^5.1.1"
anthropic = "^0.39.0"
qdrant-client = "^1.11.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
lxml = "^5.3.0"
pydantic-settings = "^2.5.2"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
numpy-financial = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
black = "^24.10.0"
ruff = "^0.7.3"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=. --cov-report=html --cov-report=term-missing"
